services/booking/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
redis==5.1.1
alembic==1.16.1
python-jose[cryptography]==3.3.0  # Для JWT decode
python-multipart==0.0.9  # Для form data позже (фото)